% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_spcs.R
\docType{data}
\name{crew_launcher_spcs}
\alias{crew_launcher_spcs}
\title{crew Launcher for SPCS}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
crew_launcher_spcs
}
\description{
R6 class that implements the crew launcher interface
for Snowflake SPCS. Each \code{launch_worker()} call creates an
EXECUTE JOB SERVICE that runs \code{mirai::daemon()} connecting back
to the controller.
}
\keyword{datasets}
