% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.check_conda_forge_availability}
\alias{.check_conda_forge_availability}
\title{Check that predict_pkgs have conda-forge counterparts}
\usage{
.check_conda_forge_availability(predict_pkgs, conda_deps)
}
\arguments{
\item{predict_pkgs}{Character vector of R package names.}

\item{conda_deps}{Character vector of conda dependencies.}
}
\description{
SPCS containers install R packages exclusively from conda-forge.
Packages installed from CRAN or GitHub in Workspace will NOT be
available at inference time.  This function warns the user if any
predict_pkgs don't appear in the resolved conda_deps.
}
\keyword{internal}
