% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{print.sfr_connection}
\alias{print.sfr_connection}
\title{Print an sfr_connection object}
\usage{
\method{print}{sfr_connection}(x, ...)
}
\arguments{
\item{x}{An \code{sfr_connection} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an sfr_connection object
}
