% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rcat}
\alias{rcat}
\title{Print text cleanly}
\usage{
rcat(...)
}
\arguments{
\item{...}{Objects to concatenate and print.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
A \code{cat()}-like function that avoids Workspace Notebook formatting issues.
}
