% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rglimpse}
\alias{rglimpse}
\title{Glimpse a data.frame}
\usage{
rglimpse(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Shows structure information similar to \code{dplyr::glimpse()} but without
requiring dplyr.
}
