% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rview}
\alias{rview}
\title{View the head of a data.frame}
\usage{
rview(x, n = 10L)
}
\arguments{
\item{x}{A data.frame.}

\item{n}{Integer. Number of rows to show. Default: 10.}
}
\value{
Invisibly returns the first \code{n} rows.
}
\description{
View the head of a data.frame
}
