% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_add_monitor_segment}
\alias{sfr_add_monitor_segment}
\title{Add a segment column to a model monitor}
\usage{
sfr_add_monitor_segment(conn, monitor_name, column)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{monitor_name}{Character.}

\item{column}{Character. Segment column name (string column in source data).}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Add a segment column to a model monitor
}
\examples{
\dontrun{
sfr_add_monitor_segment(conn, "MY_MON", "REGION")
}

}
