% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_attach_feature_desc}
\alias{sfr_attach_feature_desc}
\title{Attach per-feature descriptions}
\usage{
sfr_attach_feature_desc(fs, name, version, descs)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}

\item{descs}{A named character vector or list mapping feature names to descriptions.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Attach per-feature descriptions
}
