% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_create_dataset}
\alias{sfr_create_dataset}
\title{Create a new Snowflake Dataset}
\usage{
sfr_create_dataset(conn, name, exist_ok = FALSE)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Dataset name.}

\item{exist_ok}{Logical. If \code{FALSE} (default), errors if the dataset
already exists.}
}
\value{
An \code{sfr_dataset} object.
}
\description{
Creates an empty dataset in the current database/schema. Add versions
with \code{\link[=sfr_create_dataset_version]{sfr_create_dataset_version()}}.
}
