% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_create_image_repo}
\alias{sfr_create_image_repo}
\title{Create an image repository}
\usage{
sfr_create_image_repo(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Image repository name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Image repositories store container images for SPCS services.
}
