% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_delete_compute_pool}
\alias{sfr_delete_compute_pool}
\title{Delete a compute pool}
\usage{
sfr_delete_compute_pool(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Compute pool name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a compute pool
}
