% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_delete_dataset}
\alias{sfr_delete_dataset}
\title{Delete a dataset and all its versions}
\usage{
sfr_delete_dataset(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Dataset name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a dataset and all its versions
}
