% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_delete_dataset_version}
\alias{sfr_delete_dataset_version}
\title{Delete a dataset version}
\usage{
sfr_delete_dataset_version(conn, name, version)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Dataset name.}

\item{version}{Character. Version to delete.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a dataset version
}
