% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_delete_eai}
\alias{sfr_delete_eai}
\title{Delete an External Access Integration}
\usage{
sfr_delete_eai(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. EAI name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete an External Access Integration
}
