% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_delete_entity}
\alias{sfr_delete_entity}
\title{Delete an entity}
\usage{
sfr_delete_entity(fs, name)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Entity name to delete.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete an entity
}
