% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_delete_experiment}
\alias{sfr_delete_experiment}
\title{Delete an experiment}
\usage{
sfr_delete_experiment(exp, name = NULL)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{name}{Character. Experiment name; defaults to \code{exp$name}.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Delete an experiment
}
