% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_delete_feature_view}
\alias{sfr_delete_feature_view}
\title{Delete a Feature View}
\usage{
sfr_delete_feature_view(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to delete.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a Feature View
}
