% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_delete_image_repo}
\alias{sfr_delete_image_repo}
\title{Delete an image repository}
\usage{
sfr_delete_image_repo(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Image repository name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete an image repository
}
