% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_delete_model}
\alias{sfr_delete_model}
\title{Delete a model from the registry}
\usage{
sfr_delete_model(reg, model_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character. Name of the model to delete.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a model from the registry
}
