% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_delete_model_metric}
\alias{sfr_delete_model_metric}
\title{Delete a metric from a model version}
\usage{
sfr_delete_model_metric(reg, model_name, version_name, metric_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{metric_name}{Character.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Delete a metric from a model version
}
