% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_delete_monitor}
\alias{sfr_delete_monitor}
\title{Delete a model monitor}
\usage{
sfr_delete_monitor(reg, monitor_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{monitor_name}{Character. Monitor to delete.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Delete a model monitor
}
\examples{
\dontrun{
sfr_delete_monitor(reg, "MY_MON")
}

}
