% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_delete_run}
\alias{sfr_delete_run}
\title{Delete a run from the experiment}
\usage{
sfr_delete_run(exp, run_name)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{run_name}{Character. Name of the run to delete.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Delete a run from the experiment
}
