% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_describe_compute_pool}
\alias{sfr_describe_compute_pool}
\title{Describe a compute pool}
\usage{
sfr_describe_compute_pool(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Compute pool name.}
}
\value{
A data.frame with pool properties.
}
\description{
Describe a compute pool
}
