% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_describe_eai}
\alias{sfr_describe_eai}
\title{Describe an External Access Integration}
\usage{
sfr_describe_eai(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. EAI name.}
}
\value{
A data.frame with integration properties.
}
\description{
Describe an External Access Integration
}
