% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_describe_image_repo}
\alias{sfr_describe_image_repo}
\title{Describe an image repository}
\usage{
sfr_describe_image_repo(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Image repository name.}
}
\value{
A data.frame with repository properties.
}
\description{
Describe an image repository
}
