% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_describe_monitor}
\alias{sfr_describe_monitor}
\title{Describe a model monitor}
\usage{
sfr_describe_monitor(conn, monitor_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{monitor_name}{Character.}
}
\value{
A \code{data.frame} from \verb{DESCRIBE MODEL MONITOR}.
}
\description{
Describe a model monitor
}
\examples{
\dontrun{
sfr_describe_monitor(conn, "MY_MON")
}

}
