% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{sfr_disconnect}
\alias{sfr_disconnect}
\title{Disconnect a Snowflake connection}
\usage{
sfr_disconnect(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Closes the underlying Snowpark session.
}
