% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_end_run}
\alias{sfr_end_run}
\title{End the active experiment run}
\usage{
sfr_end_run(exp, name = NULL)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{name}{Character. Optional run name (passed to the Python SDK).}
}
\value{
Invisibly \code{TRUE}.
}
\description{
End the active experiment run
}
