% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_exp_download_artifact}
\alias{sfr_exp_download_artifact}
\title{Download artifacts for a run}
\usage{
sfr_exp_download_artifact(exp, run_name, artifact_path, target_path)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{run_name}{Character. Run name.}

\item{artifact_path}{Character. Artifact path within the run.}

\item{target_path}{Character. Local directory to write files into.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Download artifacts for a run
}
