% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_exp_list_artifacts}
\alias{sfr_exp_list_artifacts}
\title{List artifacts for a run}
\usage{
sfr_exp_list_artifacts(exp, run_name, artifact_path = NULL)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{run_name}{Character. Run name.}

\item{artifact_path}{Character. Optional prefix filter.}
}
\value{
A value from the Python SDK (typically converted by \code{reticulate}).
}
\description{
List artifacts for a run
}
