% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_exp_log_artifact}
\alias{sfr_exp_log_artifact}
\title{Log a local file as a run artifact}
\usage{
sfr_exp_log_artifact(exp, local_path, artifact_path = NULL)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{local_path}{Character. Path to the file on disk.}

\item{artifact_path}{Character. Optional path prefix inside the artifact store.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Log a local file as a run artifact
}
