% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_exp_log_metric}
\alias{sfr_exp_log_metric}
\title{Log a single metric on the active run}
\usage{
sfr_exp_log_metric(exp, key, value, step = 0L)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object.}

\item{key}{Character. Metric name.}

\item{value}{Numeric metric value.}

\item{step}{Integer step (e.g. training step or epoch).}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Log a single metric on the active run
}
