% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_export_model}
\alias{sfr_export_model}
\title{Export model files to a local directory}
\usage{
sfr_export_model(
  reg,
  model_name,
  version_name,
  target_path,
  export_mode = "model"
)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{target_path}{Character. Local directory to export to.}

\item{export_mode}{Character. Export mode (default: \code{"model"}).}
}
\value{
The target path (invisibly).
}
\description{
Export model files to a local directory
}
