% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_fv_fqn}
\alias{sfr_fv_fqn}
\title{Get the fully qualified name of a Feature View}
\usage{
sfr_fv_fqn(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}
}
\value{
A character string with the fully qualified name.
}
\description{
Get the fully qualified name of a Feature View
}
