% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_fv_lineage}
\alias{sfr_fv_lineage}
\title{Get lineage for a Feature View}
\usage{
sfr_fv_lineage(fs, name, version, direction = "both", domain_filter = NULL)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}

\item{direction}{Character. \code{"upstream"}, \code{"downstream"}, or \code{"both"}.}

\item{domain_filter}{Character vector. Optional domain filter.}
}
\value{
A data.frame or character representation of lineage.
}
\description{
Get lineage for a Feature View
}
