% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_fv_query}
\alias{sfr_fv_query}
\title{Get the underlying SQL query for a Feature View}
\usage{
sfr_fv_query(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}
}
\value{
A character string with the SQL query.
}
\description{
Get the underlying SQL query for a Feature View
}
