% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_fv_to_df}
\alias{sfr_fv_to_df}
\title{Convert a Feature View to a data.frame}
\usage{
sfr_fv_to_df(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}
}
\value{
A data.frame with the Feature View data.
}
\description{
Convert a Feature View to a data.frame
}
