% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_get_dataset}
\alias{sfr_get_dataset}
\title{Load an existing Snowflake Dataset}
\usage{
sfr_get_dataset(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Dataset name.}
}
\value{
An \code{sfr_dataset} object.
}
\description{
Load an existing Snowflake Dataset
}
