% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_get_entity}
\alias{sfr_get_entity}
\title{Get a registered entity}
\usage{
sfr_get_entity(fs, name)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Entity name.}
}
\value{
An \code{sfr_entity} object.
}
\description{
Get a registered entity
}
