% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_get_feature_view}
\alias{sfr_get_feature_view}
\title{Get a Feature View by name and version}
\usage{
sfr_get_feature_view(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to retrieve.}
}
\value{
An \code{sfr_feature_view} object.
}
\description{
Get a Feature View by name and version
}
