% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_get_model}
\alias{sfr_get_model}
\title{Get a model reference from the registry}
\usage{
sfr_get_model(reg, model_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character. Name of the model.}
}
\value{
An \code{sfr_model} object.
}
\description{
Get a model reference from the registry
}
