% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_get_model_task}
\alias{sfr_get_model_task}
\title{Get the task type of a model version}
\usage{
sfr_get_model_task(reg, model_name, version_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}
}
\value{
Character string of the task type, or \code{NULL}.
}
\description{
Get the task type of a model version
}
