% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_get_model_version}
\alias{sfr_get_model_version}
\title{Get a specific model version}
\usage{
sfr_get_model_version(reg, model_name, version_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character. Name of the model.}

\item{version_name}{Character. Version to retrieve.}
}
\value{
An \code{sfr_model_version} object.
}
\description{
Get a specific model version
}
