% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_get_monitor}
\alias{sfr_get_monitor}
\title{Get a model monitor reference}
\usage{
sfr_get_monitor(reg, name = NULL, model_name = NULL, version_name = NULL)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{name}{Character. Monitor name (if provided, \code{model_name} and
\code{version_name} are ignored).}

\item{model_name}{Character. Model name (use with \code{version_name}).}

\item{version_name}{Character. Model version name.}
}
\value{
A list with at least \code{name} and \code{status}.
}
\description{
Get a model monitor reference
}
\examples{
\dontrun{
sfr_get_monitor(reg, name = "MY_MON")
sfr_get_monitor(reg, model_name = "M", version_name = "v1")
}

}
