% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sfr_has_connection}
\alias{sfr_has_connection}
\title{Check if a Snowflake connection can be established}
\usage{
sfr_has_connection(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=sfr_connect]{sfr_connect()}}.}
}
\value{
Logical.
}
\description{
Useful for \verb{@examplesIf} and test guards.
}
