% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_list_compute_pools}
\alias{sfr_list_compute_pools}
\title{List compute pools}
\usage{
sfr_list_compute_pools(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
A data.frame.
}
\description{
List compute pools
}
