% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_list_dataset_versions}
\alias{sfr_list_dataset_versions}
\title{List versions of a dataset}
\usage{
sfr_list_dataset_versions(conn, name, detailed = FALSE)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Dataset name.}

\item{detailed}{Logical. If \code{TRUE}, returns a data.frame with full
metadata. Default: \code{FALSE} (returns character vector of version names).}
}
\value{
A character vector (default) or data.frame (if \code{detailed = TRUE}).
}
\description{
List versions of a dataset
}
