% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sfr_list_datasets}
\alias{sfr_list_datasets}
\title{List datasets in the current database/schema}
\usage{
sfr_list_datasets(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
A data.frame with dataset information.
}
\description{
List datasets in the current database/schema
}
