% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_list_eais}
\alias{sfr_list_eais}
\title{List External Access Integrations}
\usage{
sfr_list_eais(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
A data.frame.
}
\description{
List External Access Integrations
}
