% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_list_entities}
\alias{sfr_list_entities}
\title{List entities in the Feature Store}
\usage{
sfr_list_entities(fs)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}
}
\value{
A data.frame listing entities.
}
\description{
List entities in the Feature Store
}
