% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_list_feature_views}
\alias{sfr_list_feature_views}
\title{List Feature Views}
\usage{
sfr_list_feature_views(fs, entity_name = NULL, feature_view_name = NULL)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{entity_name}{Character. Filter by entity name.}

\item{feature_view_name}{Character. Filter by Feature View name.}
}
\value{
A data.frame listing Feature Views.
}
\description{
List Feature Views
}
