% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{sfr_list_fields}
\alias{sfr_list_fields}
\title{List columns/fields of a table}
\usage{
sfr_list_fields(conn, table_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{table_name}{Character. Name of the table.}
}
\value{
A data.frame with column names and types.
}
\description{
List columns/fields of a table
}
