% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_list_fv_columns}
\alias{sfr_list_fv_columns}
\title{List columns for a Feature View}
\usage{
sfr_list_fv_columns(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}
}
\value{
A data.frame with column metadata.
}
\description{
List columns for a Feature View
}
