% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_list_image_repos}
\alias{sfr_list_image_repos}
\title{List image repositories}
\usage{
sfr_list_image_repos(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
A data.frame.
}
\description{
List image repositories
}
