% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_list_services}
\alias{sfr_list_services}
\title{List services deployed for a model version}
\usage{
sfr_list_services(reg, model_name, version_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}
}
\value{
A data.frame listing services.
}
\description{
List services deployed for a model version
}
