% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{sfr_list_tables}
\alias{sfr_list_tables}
\title{List tables in the current database/schema}
\usage{
sfr_list_tables(conn, database = NULL, schema = NULL)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{database}{Character. Database to query. Default: connection's current.}

\item{schema}{Character. Schema to query. Default: connection's current.}
}
\value{
A character vector of table names.
}
\description{
List tables in the current database/schema
}
