% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_load_fvs_from_dataset}
\alias{sfr_load_fvs_from_dataset}
\title{Load Feature Views associated with a Dataset}
\usage{
sfr_load_fvs_from_dataset(fs, dataset_name, dataset_version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{dataset_name}{Character. Dataset name.}

\item{dataset_version}{Character. Dataset version.}
}
\value{
A list of lists with \code{name} and \code{version} elements.
}
\description{
Load Feature Views associated with a Dataset
}
