% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sfr_ml_version}
\alias{sfr_ml_version}
\title{Get installed snowflake-ml-python version}
\usage{
sfr_ml_version()
}
\value{
A \link{numeric_version} object, or \code{NULL} if the package is not
installed or Python is not available.
}
\description{
Get installed snowflake-ml-python version
}
\examples{
\dontrun{
sfr_ml_version()
#> [1] '1.27.0'
}

}
