% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_model_description}
\alias{sfr_model_description}
\title{Get or set model version description}
\usage{
sfr_model_description(reg, model_name, version_name, desc = NULL)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{desc}{Character. If provided, sets the description. If \code{NULL}, returns current.}
}
\value{
The description string.
}
\description{
Get or set model version description
}
