% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_model_lineage}
\alias{sfr_model_lineage}
\title{Get lineage for a model version}
\usage{
sfr_model_lineage(
  reg,
  model_name,
  version_name,
  direction = "both",
  domain_filter = NULL
)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{direction}{Character. \code{"upstream"}, \code{"downstream"}, or \code{"both"}.}

\item{domain_filter}{Character vector.}
}
\value{
Lineage information.
}
\description{
Get lineage for a model version
}
