% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_models}
\alias{sfr_models}
\title{List Model objects in the registry}
\usage{
sfr_models(reg)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}
}
\value{
A data.frame with model metadata.
}
\description{
List Model objects in the registry
}
