% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_monitor_config}
\alias{sfr_monitor_config}
\title{Model monitor registry configuration}
\usage{
sfr_monitor_config(
  model_name,
  version_name,
  function_name = "predict",
  warehouse,
  aggregation_window = "1 day"
)
}
\arguments{
\item{model_name}{Character. Registered model name.}

\item{version_name}{Character. Model version name.}

\item{function_name}{Character. Monitored prediction function name. Default:
\code{"predict"}.}

\item{warehouse}{Character. Warehouse for background monitor compute.}

\item{aggregation_window}{Character. Aggregation window, e.g. \code{"1 day"}.}
}
\value{
An object of class \code{sfr_monitor_config}.
}
\description{
Specifies the model version and warehouse used when registering a monitor
via \code{sfr_add_monitor()}.
}
\examples{
\dontrun{
cfg <- sfr_monitor_config("MY_MODEL", "v1", warehouse = "ML_WH")
}

}
