% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{sfr_read_table}
\alias{sfr_read_table}
\title{Read an entire table into a data.frame}
\usage{
sfr_read_table(conn, table_name, limit = NULL)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{table_name}{Character. Name of the table.}

\item{limit}{Integer. Maximum rows to return. Default: no limit.}
}
\value{
A data.frame.
}
\description{
Read an entire table into a data.frame
}
