% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_refresh_feature_view}
\alias{sfr_refresh_feature_view}
\title{Refresh a Feature View}
\usage{
sfr_refresh_feature_view(fs, name, version, store_type = "OFFLINE")
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to refresh.}

\item{store_type}{Character. Store to refresh: \code{"OFFLINE"} (default) or
\code{"ONLINE"}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Manually triggers a refresh of a managed Feature View.
}
