% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sfr_resume_compute_pool}
\alias{sfr_resume_compute_pool}
\title{Resume a compute pool}
\usage{
sfr_resume_compute_pool(conn, name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{name}{Character. Compute pool name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Resume a compute pool
}
