% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_resume_feature_view}
\alias{sfr_resume_feature_view}
\title{Resume a Feature View}
\usage{
sfr_resume_feature_view(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to resume.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Resumes automatic refresh of a suspended Feature View.
}
