% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_resume_monitor}
\alias{sfr_resume_monitor}
\title{Resume a model monitor}
\usage{
sfr_resume_monitor(conn, monitor_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{monitor_name}{Character.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Resume a model monitor
}
\examples{
\dontrun{
sfr_resume_monitor(conn, "MY_MON")
}

}
