% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosnowflake_queue.R
\name{sfr_scale_worker_pool}
\alias{sfr_scale_worker_pool}
\title{Scale a persistent worker pool}
\usage{
sfr_scale_worker_pool(conn, service_name, replicas)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{service_name}{Character. Name of the SPCS service.}

\item{replicas}{Integer. New number of replicas.}
}
\value{
Invisibly returns the service name.
}
\description{
Alters the MIN/MAX instances of an existing SPCS service to scale
the worker pool up or down.
}
