% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_inference.R
\name{sfr_service_endpoint}
\alias{sfr_service_endpoint}
\title{Discover the REST endpoint URL for an SPCS model service}
\usage{
sfr_service_endpoint(conn, service_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{service_name}{Character. Name of the SPCS service.}
}
\value{
An \code{sfr_endpoint} object containing the URL and metadata.
}
\description{
Queries Snowflake for the endpoint URL of a deployed SPCS service.
The returned object can be passed directly to \code{\link[=sfr_predict_rest]{sfr_predict_rest()}}.
}
\examples{
\dontrun{
ep <- sfr_service_endpoint(conn, "mpg_service")
ep
# <sfr_endpoint>
# url:     https://xyz-ak32940.snowflakecomputing.app
# service: SIMON.SCRATCH.MPG_SERVICE
}

}
