% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_set_default_model_version}
\alias{sfr_set_default_model_version}
\title{Set the default version of a model}
\usage{
sfr_set_default_model_version(reg, model_name, version_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Set the default version of a model
}
