% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_set_model_metric}
\alias{sfr_set_model_metric}
\title{Set a metric on a model version}
\usage{
sfr_set_model_metric(reg, model_name, version_name, metric_name, metric_value)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{metric_name}{Character. Name of the metric.}

\item{metric_value}{Numeric or character. Value of the metric.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Set a metric on a model version
}
