% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_show_fv_versions}
\alias{sfr_show_fv_versions}
\title{Show versions of a Feature View}
\usage{
sfr_show_fv_versions(fs, name)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}
}
\value{
A data.frame with version information.
}
\description{
Show versions of a Feature View
}
