% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_show_model_functions}
\alias{sfr_show_model_functions}
\title{List callable functions on a model version}
\usage{
sfr_show_model_functions(reg, model_name, version_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}
}
\value{
A data.frame listing functions.
}
\description{
List callable functions on a model version
}
