% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{sfr_show_model_monitors}
\alias{sfr_show_model_monitors}
\title{List model monitors in the registry}
\usage{
sfr_show_model_monitors(reg)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}
}
\value{
A \code{data.frame} of monitors (columns depend on the Python SDK).
}
\description{
List model monitors in the registry
}
\examples{
\dontrun{
sfr_show_model_monitors(reg)
}

}
