% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_show_model_versions}
\alias{sfr_show_model_versions}
\title{Show versions of a model}
\usage{
sfr_show_model_versions(reg, model_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character. Name of the model.}
}
\value{
A data.frame with version information.
}
\description{
Show versions of a model
}
