% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_show_models}
\alias{sfr_show_models}
\title{List models in the registry}
\usage{
sfr_show_models(reg)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}
}
\value{
A data.frame listing registered models.
}
\description{
List models in the registry
}
