% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_slice_feature_view}
\alias{sfr_slice_feature_view}
\title{Select a subset of features from a Feature View}
\usage{
sfr_slice_feature_view(fs, name, version, feature_names)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version.}

\item{feature_names}{Character vector of feature names to keep.}
}
\value{
An \code{sfr_feature_view} object with only the selected features.
}
\description{
Select a subset of features from a Feature View
}
