% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sfr_snowflake_version}
\alias{sfr_snowflake_version}
\title{Get Snowflake server version}
\usage{
sfr_snowflake_version(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object from \code{\link[=sfr_connect]{sfr_connect()}}.}
}
\value{
A character string with the Snowflake server version.
}
\description{
Get Snowflake server version
}
\examples{
\dontrun{
conn <- sfr_connect()
sfr_snowflake_version(conn)
#> [1] "9.26.0"
}

}
