% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sfr_start_run}
\alias{sfr_start_run}
\title{Start an experiment run}
\usage{
sfr_start_run(exp, name = NULL)
}
\arguments{
\item{exp}{An \code{sfr_experiment} object from \code{sfr_experiment()}.}

\item{name}{Character. Optional run name.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Start an experiment run
}
