% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sfr_status}
\alias{sfr_status}
\title{Check connection status}
\usage{
sfr_status(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}
}
\value{
Invisibly returns \code{TRUE} if the connection is active.
}
\description{
Check connection status
}
