% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosnowflake_queue.R
\name{sfr_stop_worker_pool}
\alias{sfr_stop_worker_pool}
\title{Stop and remove a worker pool}
\usage{
sfr_stop_worker_pool(conn, service_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{service_name}{Character. Name of the SPCS service.}
}
\value{
Invisibly returns TRUE.
}
\description{
Stop and remove a worker pool
}
