% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_suspend_feature_view}
\alias{sfr_suspend_feature_view}
\title{Suspend a Feature View}
\usage{
sfr_suspend_feature_view(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to suspend.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Pauses automatic refresh of a managed Feature View.
}
