% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{sfr_table_exists}
\alias{sfr_table_exists}
\title{Check if a table exists}
\usage{
sfr_table_exists(conn, table_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{table_name}{Character. Name of the table.}
}
\value{
Logical.
}
\description{
Check if a table exists
}
