% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sfr_undeploy_model}
\alias{sfr_undeploy_model}
\title{Undeploy a model service}
\usage{
sfr_undeploy_model(reg, model_name, version_name, service_name)
}
\arguments{
\item{reg}{An \code{sfr_model_registry} or \code{sfr_connection} object.}

\item{model_name}{Character.}

\item{version_name}{Character.}

\item{service_name}{Character.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Undeploy a model service
}
