% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_update_default_warehouse}
\alias{sfr_update_default_warehouse}
\title{Update the default warehouse for the Feature Store}
\usage{
sfr_update_default_warehouse(fs, warehouse_name)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{warehouse_name}{Character. New default warehouse name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Update the default warehouse for the Feature Store
}
