% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_update_entity}
\alias{sfr_update_entity}
\title{Update an entity's description}
\usage{
sfr_update_entity(fs, name, desc)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Entity name.}

\item{desc}{Character. New description.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Update an entity's description
}
