% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_online.R
\name{sfr_update_feature_view}
\alias{sfr_update_feature_view}
\title{Update a Feature View}
\usage{
sfr_update_feature_view(
  fs,
  name,
  version,
  refresh_freq = NULL,
  warehouse = NULL,
  desc = NULL,
  online_config = NULL
)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to update.}

\item{refresh_freq}{Character. New refresh frequency (optional).}

\item{warehouse}{Character. New warehouse (optional).}

\item{desc}{Character. New description (optional).}

\item{online_config}{An \code{\link[=sfr_online_config]{sfr_online_config()}} object (optional).}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Updates properties of an existing registered Feature View.
}
