% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sfr_use}
\alias{sfr_use}
\title{Switch warehouse, database, or schema}
\usage{
sfr_use(conn, warehouse = NULL, database = NULL, schema = NULL)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{warehouse}{Character. New warehouse name.}

\item{database}{Character. New database name.}

\item{schema}{Character. New schema name.}
}
\value{
The modified \code{sfr_connection} object (invisibly).
}
\description{
Switch warehouse, database, or schema
}
