% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosnowflake_queue.R
\name{sfr_worker_pool_status}
\alias{sfr_worker_pool_status}
\title{Get the status of a worker pool}
\usage{
sfr_worker_pool_status(conn, service_name)
}
\arguments{
\item{conn}{An \code{sfr_connection} object.}

\item{service_name}{Character. Name of the SPCS service.}
}
\value{
A list with name, status, and instance count.
}
\description{
Get the status of a worker pool
}
