% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_read_feature_view}
\alias{sfr_read_feature_view}
\title{Read data from a Feature View}
\usage{
sfr_read_feature_view(fs, name, version)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{name}{Character. Feature View name.}

\item{version}{Character. Version to read.}
}
\value{
A data.frame with feature data.
}
\description{
Read data from a Feature View
}
