% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sfr_retrieve_features}
\alias{sfr_retrieve_features}
\title{Retrieve feature values for inference}
\usage{
sfr_retrieve_features(fs, spine, features)
}
\arguments{
\item{fs}{An \code{sfr_feature_store} object.}

\item{spine}{A data.frame, SQL string, or dbplyr lazy table with entity
key values.}

\item{features}{A list of registered \code{sfr_feature_view} objects or
lists with \code{name} and \code{version}.}
}
\value{
A data.frame with feature values.
}
\description{
Retrieve feature values for inference
}
